Moving the parses to a different class
