+++ package/cloudshell/cp/aws/device_access_layer/aws_api.py
@@ -7,7 +7,8 @@ class AWSApi(object):
     def __init__(self):
         pass
 
-    def create_instance(self, ec2_session, name, ami_deployment_info):
+    @staticmethod
+    def create_instance(ec2_session, name, ami_deployment_info):
         """
         Deploys an AMI
         :param name: Will assign the deployed vm with the name
@@ -35,9 +36,10 @@ class AWSApi(object):
             # PrivateIpAddress=ami_deployment_info.private_ip_address
         )[0]
         new_name = name + ' ' + instance.instance_id
-        self.set_instance_name_and_createdby(ec2_session, instance, new_name)
 
-        # Note: pulling every 15 sec
+        TagManager.set_ami_instance_tag(ec2_session, instance, new_name)
+
+        # Note: checks every 15 sec
         instance.wait_until_running()
 
         # Reload the instance attributes
@@ -45,34 +47,34 @@ class AWSApi(object):
         return instance, new_name
 
     @staticmethod
+    def create_security_group(ec2_session, group_name, description, vpc_id):
+        return ec2_session.create_security_group(GroupName=group_name,
+                                                 Description=description,
+                                                 VpcId=vpc_id)
+
+    @staticmethod
     def get_instance_by_id(ec2_session, id):
         return ec2_session.Instance(id=id)
-
-    def set_instance_name_and_createdby(self, ec2_session, instance, name):
-        return self.set_instance_tag(ec2_session, instance, self.get_default_tags(name))
-
-    def get_default_tags(self, name):
-        return [self._get_kvp("Name", name),
-                self._get_created_by_kvp()]
 
     @staticmethod
     def create_key_pair(ec2_session, key_name):
         return ec2_session.create_key_pair(KeyName=key_name)
 
+
+class TagManager(object):
     @staticmethod
-    def set_instance_tag(ec2_session, instance, tags):
+    def set_ami_instance_tag(ec2_session, instance, tags):
         return ec2_session.create_tags(Resources=[instance.id], Tags=tags)
 
     def set_security_group_tags(self, security_group, name):
         return security_group.create_tags(Tags=self.get_default_tags(name))
+
+    def get_default_tags(self, name):
+        return [self._get_kvp("Name", name),
+                self._get_created_by_kvp()]
 
     def _get_created_by_kvp(self):
         return self._get_kvp('CreatedBy', 'Quali')
 
     def _get_kvp(self, key, value):
         return {'Key': key, 'Value': value}
-
-    def create_security_group(self, ec2_session, group_name, description, vpcid):
-        return ec2_session.create_security_group(GroupName=group_name,
-                                                 Description=description,
-                                                 VpcId=vpcid)
